third_party/WebKit/LayoutTests/screen_orientation/lockOrientation-bad-argument.html

<!DOCTYPE html>
<html>
<body>
<script src="../resources/js-test.js"></script>
<script>
description("Validates that screen.lockOrientation() returns false on bad input");
window.jsTestIsAsync = true;

function onOrientationChangeEvent(ev) {
    testFailed("Unexpected 'orientationchange' event");
}

function completeTest() {
    shouldBeEqualToString("screen.orientation", "portrait-primary");
    screen.unlockOrientation();
    finishJSTest();
}

window.addEventListener("orientationchange", onOrientationChangeEvent);

function typeErrorEnum(argStr) {
  return "TypeError: Failed to execute 'lockOrientation' on 'Screen': parameter 1 ('" + argStr + "') is not a valid enum value.";
}

shouldBeEqualToString("screen.orientation", "portrait-primary");
shouldThrow("screen.lockOrientation('invalid-orientation')", "typeErrorEnum('invalid-orientation')");
shouldBeEqualToString("screen.orientation", "portrait-primary");
shouldThrow("screen.lockOrientation(null)", "typeErrorEnum('null')");
shouldBeEqualToString("screen.orientation", "portrait-primary");
shouldThrow("screen.lockOrientation(undefined)", "typeErrorEnum('undefined')");
shouldBeEqualToString("screen.orientation", "portrait-primary");
shouldThrow("screen.lockOrientation(123)", "typeErrorEnum('123')");
shouldBeEqualToString("screen.orientation", "portrait-primary");
shouldThrow("screen.lockOrientation(window)", "typeErrorEnum('[object Window]')");
shouldBeEqualToString("screen.orientation", "portrait-primary");
shouldThrow("screen.lockOrientation(['portrait-primary', 'landscape-primary'])", "typeErrorEnum('portrait-primary,landscape-primary')");
shouldBeEqualToString("screen.orientation", "portrait-primary");
shouldThrow("screen.lockOrientation()", '"TypeError: Failed to execute \'lockOrientation\' on \'Screen\': 1 argument required, but only 0 present."');

// Finish asynchronously to give events a chance to fire.
setTimeout(completeTest, 0);
</script>
</body>
</html>
